Heiland_exe/requirements.txt

typer[all]==0.12.3
pyyaml==6.0.2
rich==13.7.1


auto-editor
